refactor logout
