+++ auth-service/routes/auth.test.ts
@@ -53,143 +53,6 @@ describe("Auth Routes", () => {
     vi.clearAllMocks();
   });
 
-  describe("POST /auth/login", () => {
-    it("should login successfully", async () => {
-      const userData = {
-        email: "test@example.com",
-        password: "password123",
-      };
-
-      const user = {
-        id: 1,
-        email: userData.email,
-        createdAt: new Date(),
-        lastLoginAt: null,
-      };
-
-      const authData = {
-        userId: user.id,
-        email: user.email,
-        passwordHash: Buffer.from("hashed-password"),
-        verifiedAt: null,
-        verificationToken: null,
-        verificationTokenExpiresAt: null,
-        forgotPasswordToken: null,
-        forgotPasswordTokenExpiresAt: null,
-      };
-
-      vi.spyOn(db.users, "getByEmail").mockResolvedValue(user);
-      vi.spyOn(db.emailAuth, "getByEmail").mockResolvedValue(authData);
-      vi.spyOn(db.users, "updateLastLogin").mockResolvedValue({
-        ...user,
-        lastLoginAt: new Date(),
-      });
-      vi.spyOn(db.users, "getById").mockResolvedValue(user);
-
-      const response = await request(app).post("/auth/login").send(userData);
-
-      expect(response.status).toBe(200);
-      expect(response.body).toMatchObject({
-        email: userData.email,
-        id: user.id,
-      });
-      expect(response.header["set-cookie"]).toBeDefined();
-
-      expect(db.users.getByEmail).toHaveBeenCalledWith(userData.email);
-      expect(db.emailAuth.getByEmail).toHaveBeenCalledWith(userData.email);
-      expect(argon2.verify).toHaveBeenCalled();
-      expect(db.users.updateLastLogin).toHaveBeenCalledWith(user.id);
-    });
-
-    it("should return 401 for invalid credentials", async () => {
-      const userData = {
-        email: "test@example.com",
-        password: "wrong",
-      };
-
-      vi.spyOn(db.users, "getByEmail").mockRejectedValue(
-        new db.users.UserNotFoundError(),
-      );
-
-      const response = await request(app).post("/auth/login").send(userData);
-
-      expect(response.status).toBe(401);
-      expect(response.body).toEqual({ message: "Invalid credentials" });
-    });
-
-    it("should return 401 for wrong password", async () => {
-      const userData = {
-        email: "test@example.com",
-        password: "wrong",
-      };
-
-      const user = {
-        id: 1,
-        email: userData.email,
-        createdAt: new Date(),
-        lastLoginAt: null,
-      };
-
-      const authData = {
-        userId: user.id,
-        email: user.email,
-        passwordHash: Buffer.from("hashed-password"),
-        verifiedAt: null,
-        verificationToken: null,
-        verificationTokenExpiresAt: null,
-        forgotPasswordToken: null,
-        forgotPasswordTokenExpiresAt: null,
-      };
-
-      vi.spyOn(db.users, "getByEmail").mockResolvedValue(user);
-      vi.spyOn(db.emailAuth, "getByEmail").mockResolvedValue(authData);
-      vi.spyOn(argon2, "verify").mockResolvedValue(false);
-
-      const response = await request(app).post("/auth/login").send(userData);
-
-      expect(response.status).toBe(401);
-      expect(response.body).toEqual({ message: "Invalid credentials" });
-    });
-  });
-
-  describe("POST /auth/logout", () => {
-    it("should logout successfully", async () => {
-      // First login
-      const userData = {
-        email: "test@example.com",
-        password: "password123",
-      };
-
-      const user = {
-        id: 1,
-        email: userData.email,
-        createdAt: new Date(),
-        lastLoginAt: null,
-      };
-
-      vi.spyOn(db.users, "getByEmail").mockResolvedValue(user);
-      vi.spyOn(db.emailAuth, "getByEmail").mockResolvedValue({
-        userId: user.id,
-        email: user.email,
-        passwordHash: Buffer.from("hashed-password"),
-        verifiedAt: null,
-        verificationToken: null,
-        verificationTokenExpiresAt: null,
-        forgotPasswordToken: null,
-        forgotPasswordTokenExpiresAt: null,
-      });
-
-      // Login first
-      await request(app).post("/auth/login").send(userData);
-
-      // Then logout
-      const response = await request(app).post("/auth/logout");
-
-      expect(response.status).toBe(200);
-      expect(response.body).toEqual({});
-    });
-  });
-
   describe("GET /auth/verify", () => {
     it("should return 401 when not authenticated", async () => {
       const response = await request(app).get("/auth/verify");
